Refactor: modulize the original lib/marked.js
