+++ lib/renderText.js
@@ -1,0 +1,23 @@
+/**
+ * Text renderer
+ */
+
+class TextRenderer {
+  constructor() {
+
+  }
+
+  text() {
+    return text
+  }
+
+  image(href, title, text) {
+    return '' + text
+  }
+
+  br() {
+    return ''
+  }
+}
+
+export default TextRenderer
